Añadida funcionalidad  filtrar pelicula en ventana
Se implementó el filtrado de películas en la ventana de mostrar catálogo

+++ src/main/java/com/mycompany/videoclub/proyecto/Gestor.java
@@ -279,7 +279,7 @@ public class Gestor{
     }
 
       //Se Realiza una busqueda de peliculas que tenga rating mayor o igual al ingresado por el usuario y se devuelve una lista con las peliculas que cumplen con el criterio
-      public ArrayList<Pelicula> buscarPeliculasPorRating(float rating){
+      public ArrayList<Pelicula> buscarPeliculasPorRating(double rating){
 
         ArrayList<Pelicula> listaFiltrada = new ArrayList<>();
